fixed bugs, renamed aw-sunburst to aw-sunburst-clock
